video 配信の部分を responseVideoFile へ分けた
