+++ nodejs/route.js
@@ -27,6 +27,8 @@ function route(parsedUrl, response, request, postData) {
     log.access.info("About to route a request for " + parsedUrl.pathname);
     if(typeof handle[parsedUrl.pathname] == 'function') {
         return handle[parsedUrl.pathname](response, parsedUrl, request, postData);
+    } else if(parsedUrl.pathname.split(path.sep)[1] == "video") {
+        return requestHandlers.responseVideoFile(response, request, parsedUrl, fileTypeHash);
     } else if (path.extname(parsedUrl.pathname) in fileTypeHash) {
         return requestHandlers.responseSpecifiedFile(response, request, parsedUrl, fileTypeHash);
     } else {
